S10 p31 a103 110 싱글 플레이 결과 등록 로직 작성

See merge request s10-final/S10P31A103!98

+++ backend/src/main/java/com/vita/backend/character/data/response/detail/CharacterGameSingleRankingDetail.java
@@ -9,6 +9,6 @@ public record CharacterGameSingleRankingDetail(
 	@JsonProperty("nickname")
 	String nickname,
 	@JsonProperty("score")
-	Double score
+	Long score
 ) {
 }
